Move _safetensors module to integrations/
The safetensors loader is an integration with the HuggingFace/safetensors
wire format — it belongs alongside gguf and ort_genai in the integrations
package rather than as a top-level private module.

- git mv src/mobius/_safetensors.py → src/mobius/integrations/_safetensors.py
- git mv src/mobius/_safetensors_test.py → src/mobius/integrations/_safetensors_test.py
- Update all three call sites to import from mobius.integrations._safetensors:
  _weight_loading.py, __main__.py, _diffusers_builder.py
